Move adapter files to the adapter package
