+++ app/src/main/java/com/github/polybooks/activities/ListBooksActivity.kt
@@ -4,7 +4,7 @@ import android.content.Intent
 import androidx.recyclerview.widget.RecyclerView
 import com.github.polybooks.R
 import com.github.polybooks.core.Book
-import com.github.polybooks.core.database.BooksAdapter
+import com.github.polybooks.adapter.database.BooksAdapter
 import com.github.polybooks.database.BookSettings
 import com.github.polybooks.database.Query
 import com.github.polybooks.utils.setupNavbar
