+++ app/src/main/java/com/github/polybooks/activities/ListSalesActivity.kt
@@ -7,7 +7,7 @@ import com.github.polybooks.core.Sale
 import com.github.polybooks.core.SaleState
 import com.github.polybooks.database.Query
 import com.github.polybooks.database.SaleSettings
-import com.github.polybooks.database.SalesAdapter
+import com.github.polybooks.adapter.database.SalesAdapter
 import com.github.polybooks.utils.setupNavbar
 import com.google.android.material.bottomnavigation.BottomNavigationView
